Implemented mock of MarkdownView and Editor, and got several example tests working.
