+++ __tests__/paste-text.tests.ts
@@ -1,0 +1,23 @@
+import { MarkdownView } from '../__mocks__/obsidian';
+
+import { pasteText } from '../src/paste-text';
+
+// describe('Examining toggle-blockquote-at-current-indentation', () => {
+//     beforeAll(() => {
+//       const view = new MarkdownView(
+//         [
+//           'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
+//           'Sed venenatis lectus et leo viverra, ac viverra purus rutrum.',
+//           '',
+//           'Etiam semper massa ut est faucibus, eu luctus arcu porttitor.'
+//         ],
+//         {line: 0, ch: 0},
+//         {line: 0, ch: 0},
+//         {line: 0, ch: 0}
+//       );
+//     });
+
+//     test('Adds blockquote to single line', () => {
+//       return
+//     })
+// });
