+++ jest.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  transform: {
+    "^.+\\.ts?$": "ts-jest",
+  },
+  testEnvironment: "jsdom",
+  // testRegex: "__tests__/.*\\.test?\\.ts$",
+  testMatch: ["**/__tests__/*.ts"],
+  moduleFileExtensions: ["ts", "js"],
+};
