+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "dev": "rollup --config rollup.config.js -w",
     "build": "rollup --config rollup.config.js --environment BUILD:production",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest --passWithNoTests",
+    "test-dev": "jest --passWithNoTests --watch --verbose --silent=false '__tests__/toggle-quote.ts'"
   },
   "keywords": [],
   "author": "Jacob Levernier",
@@ -15,6 +16,7 @@
     "@rollup/plugin-commonjs": "^18.0.0",
     "@rollup/plugin-node-resolve": "^11.2.1",
     "@rollup/plugin-typescript": "^8.2.1",
+    "@types/jest": "^26.0.24",
     "@types/node": "^14.14.37",
     "chai": "^4.3.4",
     "jest": "^27.0.6",
@@ -22,8 +24,10 @@
     "obsidian": "^0.12.0",
     "rollup": "^2.32.1",
     "spectron": "^15.0.0",
+    "ts-jest": "^27.0.3",
     "tslib": "^2.2.0",
-    "typescript": "^4.2.4"
+    "tslint": "^6.1.3",
+    "typescript": "^4.3.5"
   },
   "dependencies": {}
 }
