+++ tsconfig.json
@@ -9,6 +9,8 @@
     "noImplicitAny": true,
     "moduleResolution": "node",
     "importHelpers": true,
+    "esModuleInterop": true,
+    "types": ["node", "jest"],
     "lib": [
       "dom",
       "es5",
